Add hacky support for Switch Mii colors
* Add "nnmiiCommonColors", "nnmiiFacelineColors" tables into
FFLiColor.cpp.
NOTE: these are NOT actually from reversing nn::mii, they are from the
Mii Studio JS because I could not find the floats yet.
* Add (hacky) "markCommonColor/unmarkCommonColor" functions to indicate
whether a color is supposed to be grabbed from the Switch
common/faceline color tables.
Which takes advantage of the fact that FFLiCharInfo color fields are
s32s.