Add ABSURD hack to allow new glass types
Hoooly crap this is such a terrible hack buuuut I also don't want to
rewrite it to use templates heheheheee LOOOO

... So, I added some code to FFLResource.py that will detect 20 glass
types and expand the resource header if they are found. I also expanded
it in FFL itself. However, the problem comes when you do that and then
try to use old assets.
What I did is some hack to calculate the offset of EVERYTHING AFTER THE
GLASS TYPE ARRAY, which would be a negative asset for an old resource,
and there ya go, it actually does work even on a Wii U. (I think)
As you can see, we differentiate the resources by the version. Did I
bump up the version? Nope, but I put the amount of glass types in the
first byte and size of the texture header in the last two bytes - that's
used to calculate where the shape header is.

I made this script "texturec from aflreshigh.sh" to import higher
quality textures from AFLResHigh.dat - but NOT AFLResHigh_2_3.dat!!!!! -
into FFLResHigh.dat, though I'm really not intending it to be used by
anyone, it doesn't even convert the glass types because I can't find a
DDS encoder that converts to A8L8 or RG8 at the moment other than
GIMP's, so I did all of those by hand. Texturec did most of the DDSes
but not those, ehhhhhhhhhh