Move expression flags to u64s (60 maximum)
In this commit I attempted to overcome the limit of only being able to
access 31 out of Miitomo's 70 expression flags by making expressionFlag
a u64. It didn't take much effort and, this still isn't a real solution
due to it only getting about 60 of them in, but I'm pushing it anyway.
Expect me to revert this, or something.