Make expression flag type into a typedef
This is so that you can make it a 128-bit int if you can, and that's the
path of least resistance to supporting all expressions, which this now
will with this commit.
Also fixes some static asserts on 32 bit, and readds the expressionj
flag check when checking the model descriptor.