Make resource header type polymorphic, support FFL alongside AFLResHigh_2_3 types
* This commit makes FFLiResourceHeader abstract, meaning it is accessed
exclusively by methods now, and the classes implementing it are
FFLiResourceHeaderDefault for FFLResHigh.dat/FFLResMiddle.dat, and
FFLiResourceHeaderAFL_2_3 for AFLResHigh_2_3.dat (standard
AFLResHigh.dat not supported, though it could be)

* Little endian resources are now supported, by checking if the magic is
backwards from what it should be (not tested on Wii U yet)

* Two methods determine the type of resource:
  - The total uncompressed size, which is _c
- The first three bits of the total uncompressed size, which
FFLResource.py now injects for FFL, AFL, AFL 2.3.0
- Note that FFL and AFL never check this field. Also, I can't
reproduce nearly the same value  as this in FFLResource.py but won't be
looking into it further

* FFLiGetTextureResoucePartsInfos and FFLiGetTextureResourceNum now have
different arguments, though these should only be used internally so
shouldn't be much of a problem.

* NOTE: Loading the default resource paths/as files without specifying
the resources themselves in FFLInitResEx is NOT SUPPORTED right now,
since I've had trouble with it.