Add Miitomo "blank" expression
Assuming that 61 and 62 are those expressions, and it does the same
thing the cat/dog expressions are doing and simply removes the mask
shape. In hindsight, I wish this was just disablable and using
FFLiSetExpression would simply hide either the nose/noseline, or mask,
depending on the expression. Maybe that's worth correcting later.