Add FFLSetFrontCullForFlipX, FFLiSetNormalIsSnorm8_8_8_8 in place of definition
FFLiSetNormalIsSnorm8_8_8_8(bool) is a function you can call instead of
FFL_NORMAL_ATTRIBUTE_IS_SNORM_8_8_8_8, to enable FFL to convert normals
from R10G10B10A2_SNORM to R8G8B8A8_SNORM, a format that is more
compatible, particularly with OpenGL ES 2.0 and, also it looks like
Veldrid doesn't support it either...?????

Similarly, FFLSetFrontCullForFlipX(bool), when set to false, will adjust
the index buffer of flipped hair so that you can continue to use
backface culling. This is basically what nn::mii is doing.

Decided to put both of those in an enable function so that if FFL is
built as a library, they can be enabled and disabled at will per
configuration. Will probably be doing this for other things as well...
texture callback? No render texture?