Fix FFLResource.py bug with LE causing huge allocations, other misc logging
* Fix a critical bug in FFLResource.py when packing resources in little
endian (-LE) where partsMaxSize for the texture header was still in big
endian and were interpreted as huge numbers. These partsMaxSize entries
are passed to rio::MemUtil::alloc, so as a result this bug caused huge
memory allocations but only in between FFLInitCharModelCPUStep and
GPUStep, aargghhh.
- Maybe not as critical since I was the only one using this feature,
though it seems to be the reason Emscripten demos weren't working on my
phone...
- But as a result, I have also added asserts in cases where those
values are over 200 MB large. It's not implausible that they could be
that big, but just a sanity check.

* A stub for FFLiEnableSpecialMii was added if you have
FFL_NO_DATABASE_FILE defined so that it can still be called.
* Another case for FFL_RESULT_ERROR now has logging when an invalid
resource type is passed in.
* Forgot to account for FFL_NO_NINTEXUTILS case when using texture
callback in my last commit, argh.

* More detailed logging has been added to
FFLiDatabaseFileAccessor::BootLoad when database files fail to load.