Remove "opengl" build and associated code to simplify things
This removes the build target that includes a portion of RIO so that the
library can still upload textures and draw faceline/masks when embedding
FFL in OpenGL applications.

Out of all of the things I've made with FFL: FFLSharp, FFL.js... only
the ffl-raylib-samples used this. It was clunky needing to initialize
its GLAD GL context, and the RIO/C++ dependency was still there...
Because I'm about to rewrite the raylib samples to use rlgl.h, which
abstracts to GL 3.3/2.1/ES2/1.1..., I thought I'd remove this. Most
people probably want Miis to work by importing the model or drawing in
something that isn't using exclusively OpenGL.