Simplify mask code with equivalents
Most of this is negating the use of resolution in FFLiRawMask. While
CalcRawMask takes resolution and uses it to make "baseScale", and it's
also used to make the ortho projection... in reality, if you hardcode
it, it achieves the exact same goal and I can't notice a single
difference. The resolution here is actually irrelevant to the render
texture's resolution itself.
Two nuances here:

* You can see FFLiiGetAdjustedEyeH/FFLiiGetAdjustedMouthH were commented
out, due to them causing some unintended distortion in mouth/eye height
(Y-scale) if baseScale is 1.0.
I have seen that nn::mii does the same, but removes the code inside the
function. Perhaps it never did anything useful.
* I've replaced the function to get the ortho matrix with a custom one
that does a "good enough" job.
As well as that, I added some code to CalcMVMatrix that can build it
into the MV matrix itself. So the fourth row of the matrix isn't needed
in this case.
This should get us closer to making all mask transforms into pure affine
SRT transformations, which would just make things easier instead of
using matrices.